# dc-api/openapi.yaml
openapi: 3.0.3
info:
  title: DC-API
  version: 0.1.0
  description: |
    WSO2 Sovereign Cloud Control Plane API.

    DC-API sits in front of a Rancher + Harvester infrastructure layer and
    exposes a clean, tenant-isolated REST API for managing VMs, RKE2 clusters,
    virtual networks, DNS zones, and RBAC — in the style of AWS / Azure.

    **Authentication**

    Every `/v1/*` endpoint requires a `Bearer` token in the `Authorization`
    header. Two token types are accepted by the same scheme:

    1. **Asgardeo OIDC JWT** — obtained via the `dcctl login` PKCE flow. The JWT
       must carry a `groups` claim containing at least one group matching the
       tenant group prefix (default `dc-tenant-<tenant-id>`).

    2. **Service-account token** — a long-lived opaque token with the format
       `dcapi_sa_<lookup_id>_<secret>`, issued by
       `POST /v1/tenants/{tenant_id}/service-accounts`. Store it on creation;
       it cannot be retrieved again.

    **Async provisioning**

    Long-running operations (VM create, VNet create, etc.) return `202 Accepted`
    immediately. The response body contains the resource object with
    `"status": "PENDING"` and a `note` field with the polling URL. Poll the
    GET endpoint until `status` transitions to `ACTIVE` or `FAILED`.

    **Status lifecycle**

    `PENDING → ACTIVE` or `PENDING → FAILED`

    Deletion: `ACTIVE → DELETING → (row removed on success)`

    **Error format**

...
servers:
...
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /v1/tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List tenants the authenticated principal has access to
      description: |
        Returns the set of tenants for which the authenticated principal holds
        at least one `role_assignments` row (at any scope within the tenant).

        The `roles` array is the **union of distinct roles** the principal holds
        across all scopes within that tenant — useful for the UI tenant switcher
        to decide which actions to enable.

        Both human users (OIDC JWT) and service accounts (dcapi_sa_* token)
        may call this endpoint. The result is always scoped to the caller's
        own principal.

        **Platform admins** (sub in `DCAPI_PLATFORM_ADMIN_SUBS`, OR member of
        `DCAPI_ADMIN_GROUP`): this endpoint returns **every tenant** in the
        `tenants` registry — not just the caller's role-assignment rows.
        Each entry's `roles` is set to `["owner"]` for UI convenience;
        admins bypass per-tenant RBAC checks at the middleware layer
        regardless. Tenants without role-assignment rows are still
        visible to admins as long as they're in the registry — the
        registry is populated by `POST /v1/admin/tenants` (explicit
        admin pre-registration) or by `POST /v1/tenants/{tid}/members`
        (UPSERT on first invite).

        **Required auth**: any valid token. No minimum role required beyond
        being authenticated.
      responses:
        "200":
          description: List of tenants the principal has access to
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantSummary"
              example:
                - id: iam-platform
                  name: iam-platform
                  roles:
                    - owner
                - id: dev-team
                  name: dev-team
                  roles:
                    - member
                    - viewer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Admin: tenant registry ────────────────────────────────────────────────────
...
  /v1/tenants/{tenant_id}/projects/{project_id}/keyvaults:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"

    post:
      operationId: createKeyVault
      tags: [keyvaults]
      summary: Create a Key Vault (synchronous)
      description: |
        Creates a Key Vault as a logical container for secrets. Synchronous —
        returns `201 Created` with status `ACTIVE`.

        Chunk 1 stores only the vault metadata (name, tenant ownership, soft-
        delete window). Later chunks add per-VPC Private Endpoints, access
        policies, and the OpenBao-backed secret data plane.

        **Required role**: `member` or `owner`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeyVaultRequest"
            example:
              name: billing-secrets
              soft_delete_days: 30
      responses:
        "201":
          description: Key Vault created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyVault"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listKeyVaults
      tags: [keyvaults]
      summary: List Key Vaults for the authenticated tenant
      responses:
        "200":
          description: List of Key Vaults
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KeyVault"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/tenants/{tenant_id}/projects/{project_id}/keyvaults/{id}:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"
      - name: id
        in: path
        required: true
        description: UUID of the Key Vault
        schema:
          type: string
          format: uuid
          example: 6f1c4d2a-0000-0000-0000-000000000071

    get:
      operationId: getKeyVault
      tags: [keyvaults]
      summary: Get a Key Vault by UUID
      responses:
        "200":
          description: Key Vault resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyVault"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteKeyVault
      tags: [keyvaults]
      summary: Delete a Key Vault
      description: |
        Hard-deletes the Key Vault row. Chunk 1 has no secret data plane, so
        this is immediate. When the OpenBao mount lands the same endpoint
        becomes a soft delete + reconciler-driven teardown.

        **Required role**: `member` or `owner`.
      responses:
        "204":
          description: Key Vault deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/tenants/{tenant_id}/projects/{project_id}/keyvaults/{id}/secrets:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"
      - name: id
        in: path
        required: true
        description: UUID of the parent Key Vault
        schema:
          type: string
          format: uuid
          example: 6f1c4d2a-0000-0000-0000-000000000071

    get:
      operationId: listKeyVaultSecrets
      tags: [keyvaults]
      summary: List secret names in a Key Vault (no values returned)
      description: |
        Returns a page of `KeyVaultSecretSummary` objects. Only secret **names**,
        version numbers, and timestamps are returned — values are never included
        in list responses.

        Soft-deleted secrets (with a non-null `deleted_at`) are included in the
        list so callers can see what is pending purge and still within the vault's
        `soft_delete_days` window.

        Cursor-based pagination: supply `?cursor=<opaque>` from the previous
        response's `next_cursor` field. Omit for the first page.

        **Required role**: `viewer`, `member`, or `owner`.
      parameters:
        - name: cursor
          in: query
          required: false
          description: |
            Opaque pagination cursor returned by the previous response's
            `next_cursor` field. Omit to start from the first page.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: |
            Maximum number of secrets to return per page. Default 100; max 500.
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Paginated list of secret name summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyVaultSecretListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Accepts either:
        1. An Asgardeo OIDC JWT (RS256) — obtained via the PKCE flow (`dcctl login`).
        2. A DC-API service-account token with the format
           `dcapi_sa_<lookup_id>_<secret>` — issued by
           `POST /v1/tenants/{tenant_id}/service-accounts`.
  parameters:
...
    TenantID:
      name: tenant_id
      in: path
      required: true
      description: Tenant identifier (slug, e.g. "iam-platform")
      schema:
        type: string
        example: iam-platform
    ProjectID:
      name: project_id
      in: path
      required: true
      description: Project identifier (slug, e.g. "platform-infra")
      schema:
        type: string
        example: platform-infra
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "missing Authorization header"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "failed to register resource"
...
  schemas:
...
    TenantSummary:
      type: object
      required: [id, name, roles]
      description: |
        A tenant the authenticated principal has access to, along with
        the union of roles the principal holds across all scopes within
        that tenant.
      properties:
        id:
          type: string
          description: Tenant slug identifier
          example: iam-platform
        name:
          type: string
          description: Display name (same as id in M1.5 — separate name field is M5)
          example: iam-platform
        roles:
          type: array
          items:
            type: string
            enum: [owner, member, viewer]
          description: |
            Union of distinct roles the principal holds across all scopes
            within this tenant. Used by the UI to decide which actions to enable.
          example:
            - owner
...
    CreateKeyVaultRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: |
            Vault name. Must be a valid DNS label (lowercase alphanumerics and
            hyphens, 3..63 chars, starts with a letter). Uniqueness enforced
            per tenant.
          example: billing-secrets
        soft_delete_days:
          type: integer
          minimum: 7
          maximum: 90
          default: 30
          description: |
            Recovery window for soft-deleted secrets, in days. Range 7..90;
            omit or set to `0` to use the default of `30`. Chunk 1 stores
            the value but does not yet act on it — chunk 3 (secret delete /
            recovery) will enforce the window.
          example: 30
...